Bug# 1614 Changed binning and bead diameter type to int from long.

+++ Etomo/src/etomo/ApplicationManager.java
@@ -2384,8 +2384,8 @@ public final class ApplicationManager extends BaseManager implements
     return retval;
   }
 
-  public long getBeadfixerDiameter(AxisID axisID) {
-    return Math.round(metaData.getFiducialDiameter() / metaData.getPixelSize()
+  public int getBeadfixerDiameter(AxisID axisID) {
+    return Math.round((float) (metaData.getFiducialDiameter() / metaData.getPixelSize())
         / UIExpertUtilities.INSTANCE.getStackBinning(this, axisID, ".preali"));
   }
 
@@ -4152,7 +4152,7 @@ public final class ApplicationManager extends BaseManager implements
   }
 
   public boolean equalsBinning(AxisID axisID, int binning, FileType fileType) {
-    long fileBinning = UIExpertUtilities.INSTANCE.getStackBinning(this, axisID, fileType);
+    int fileBinning = UIExpertUtilities.INSTANCE.getStackBinning(this, axisID, fileType);
     return binning == fileBinning;
   }
